apps/agents/pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = "RowBoat Labs Agent OS"
authors = ["Akhilesh <akhilesh@rowboatlabs.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/rowboatlabs/agents"
package-mode = false

[tool.poetry.dependencies]
# Python
python = ">=3.10,<4.0"

# Dependencies
annotated-types = "^0.7.0"
anyio = "^4.6.2"
beautifulsoup4 = "^4.12.3"
blinker = "^1.8.2"
certifi = "^2024.8.30"
charset-normalizer = "^3.4.0"
click = "^8.1.7"
distro = "^1.9.0"
dnspython = "^2.7.0"
et_xmlfile = "^2.0.0"
eval_type_backport = "^0.2.0"
firecrawl = "^1.4.0"
Flask = "^3.0.3"
h11 = "^0.14.0"
httpcore = "^1.0.6"
httpx = "^0.27.2"
idna = "^3.10"
itsdangerous = "^2.2.0"
Jinja2 = "^3.1.4"
jiter = "^0.6.1"
jsonpath-python = "^1.0.6"
lxml = "^5.3.0"
markdownify = "^0.13.1"
MarkupSafe = "^3.0.2"
mypy-extensions = "^1.0.0"
nest-asyncio = "^1.6.0"
numpy = "^2.1.2"
openai = "^1.52.2"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
pydantic = "^2.9.2"
pydantic_core = "^2.23.4"
pymongo = "^4.10.1"
python-dateutil = "^2.8.2"
python-docx = "^1.1.2"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
requests = "^2.32.3"
setuptools = "^75.1.0"
six = "^1.16.0"
sniffio = "^1.3.1"
soupsieve = "^2.6"
tabulate = "^0.9.0"
tqdm = "^4.66.5"
typing-inspect = "^0.9.0"
typing_extensions = "^4.12.2"
tzdata = "^2024.2"
urllib3 = "^2.2.3"
websockets = "^13.1"
Werkzeug = "^3.0.5"
wheel = "^0.44.0"